backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
et-xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.124.4
h11==0.16.0
idna==3.11
motor==3.6.1
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic-core==2.27.2
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
resend==2.19.0
rsa==4.9.1
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
uvicorn==0.33.0
